+++ src/Vouchers.php
@@ -79,6 +79,9 @@ class Vouchers
         }
 
         if (isset($attributes['redirect'])) {
+            if (!filter_var($attributes['redirect'], FILTER_VALIDATE_URL)) {
+                return __('Error: redirect is set, but is an invalid URL', Plugin::TEXT_DOMAIN);
+            }
             $extraOptions[] = "redirect_url: '" . $attributes['redirect'] . "'";
         }